backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.0
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-core==0.3.25
langchain-community==0.3.7
anthropic==0.40.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.12
sse-starlette==2.1.3
pandasql==0.7.3
pandas==2.2.3
pyarrow==18.1.0
jira==3.8.0